# api/templates/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MetaTrader 5 Flask REST API",
    "description": "Comprehensive REST API exposing 100% of the official Python MetaTrader 5 library methods in a headless Wine container.\n\n### Modules\n- **System**: MT5 lifecycle, connection, login, terminal and account status.\n- **Symbols & DOM**: Symbol listings, specifications, real-time ticks, and Depth of Market.\n- **Market Data**: Historical rates (OHLCV) and ticks with flexible filtering.\n- **Trading**: Market/pending order execution, positions management, margin & profit calculators.\n- **History**: Closed orders and deals history with date and ticket filters.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Local API Server"
    }
  ],
...
  "paths": {
...
    "/api/order/open": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Open Market or Pending Order",
        "description": "Simplified high-level endpoint to place BUY, SELL, or pending limit/stop orders with SL and TP.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderOpenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order execution result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or execution error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order/close": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Close Position",
        "description": "Closes an open position fully or partially by ticket ID.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCloseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position closure result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order/modify": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Modify Position SL/TP",
        "description": "Updates the Stop Loss and Take Profit levels of an open position.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderModifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position modification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order/{ticket}": {
      "delete": {
        "tags": [
          "Trade"
        ],
        "summary": "Cancel Pending Order",
        "description": "Deletes/cancels an active pending order by its ticket ID.",
        "parameters": [
          {
            "name": "ticket",
            "in": "path",
            "required": true,
            "description": "Order ticket ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order_send": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Send Raw MT5 Order Request",
        "description": "Directly sends an official MqlTradeRequest dictionary to MT5 via `order_send`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSendRawRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Raw order result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order_check": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Check Order Pre-Execution",
        "description": "Validates account margin, funds, and parameters for an order without sending it to the market.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSendRawRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order check result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order_calc_margin": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Calculate Required Margin",
        "description": "Calculates the margin required to open a trade of specified type, symbol, volume, and price.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCalcMarginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calculated margin amount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order_calc_profit": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Calculate Expected Profit",
        "description": "Calculates profit/loss in account currency for specified opening and closing prices.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCalcProfitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calculated profit amount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/positions_get": {
      "get": {
        "tags": [
          "Trade"
        ],
        "summary": "Get Open Positions List",
        "description": "Returns array of active open positions. Can filter by symbol, group pattern, or position ticket.",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Open positions list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders_get": {
      "get": {
        "tags": [
          "Trade"
        ],
        "summary": "Get Active Pending Orders List",
        "description": "Returns list of active pending orders. Can filter by symbol, group, or ticket.",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "OrderOpenRequest": {
        "type": "object",
        "required": [
          "symbol",
          "volume"
        ],
        "properties": {
          "symbol": {
            "type": "string",
            "example": "EURUSD"
          },
          "type": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL",
              "BUY_LIMIT",
              "SELL_LIMIT",
              "BUY_STOP",
              "SELL_STOP",
              "BUY_STOP_LIMIT",
              "SELL_STOP_LIMIT"
            ],
            "default": "BUY",
            "description": "Order type string"
          },
          "volume": {
            "type": "number",
            "example": 0.01
          },
          "price": {
            "type": "number",
            "description": "Required for limit/stop orders. Optional for market orders (defaults to current ask/bid)",
            "example": 1.085
          },
          "sl": {
            "type": "number",
            "description": "Stop Loss price level",
            "example": 1.08
          },
          "tp": {
            "type": "number",
            "description": "Take Profit price level",
            "example": 1.095
          },
          "deviation": {
            "type": "integer",
            "default": 20,
            "description": "Maximum price slippage in points"
          },
          "comment": {
            "type": "string",
            "default": "API Order"
          },
          "magic": {
            "type": "integer",
            "default": 0,
            "description": "Expert Advisor Magic Number ID"
          },
          "type_filling": {
            "type": "integer",
            "description": "Order execution filling mode (0 = FOK, 1 = IOC, 2 = RETURN, 3 = BOC)",
            "example": 1
          }
        }
      },
      "OrderCloseRequest": {
        "type": "object",
        "required": [
          "ticket"
        ],
        "properties": {
          "ticket": {
            "type": "integer",
            "description": "Position Ticket ID to close",
            "example": 1234567
          },
          "volume": {
            "type": "number",
            "description": "Volume to close (default: closes entire position volume)",
            "example": 0.01
          },
          "deviation": {
            "type": "integer",
            "default": 20
          },
          "comment": {
            "type": "string",
            "default": "API Close"
          }
        }
      },
      "OrderModifyRequest": {
        "type": "object",
        "required": [
          "ticket"
        ],
        "properties": {
          "ticket": {
            "type": "integer",
            "description": "Position Ticket ID",
            "example": 1234567
          },
          "sl": {
            "type": "number",
            "description": "New Stop Loss level",
            "example": 1.082
          },
          "tp": {
            "type": "number",
            "description": "New Take Profit level",
            "example": 1.098
          }
        }
      },
      "OrderSendRawRequest": {
        "type": "object",
        "required": [
          "action",
          "symbol",
          "volume",
          "type"
        ],
        "properties": {
          "action": {
            "type": "integer",
            "description": "1 = TRADE_ACTION_DEAL, 5 = TRADE_ACTION_PENDING, 6 = TRADE_ACTION_SLTP, 8 = TRADE_ACTION_REMOVE",
            "example": 1
          },
          "symbol": {
            "type": "string",
            "example": "EURUSD"
          },
          "volume": {
            "type": "number",
            "example": 0.1
          },
          "type": {
            "type": "integer",
            "description": "0 = ORDER_TYPE_BUY, 1 = ORDER_TYPE_SELL",
            "example": 0
          },
          "price": {
            "type": "number",
            "example": 1.085
          },
          "sl": {
            "type": "number",
            "example": 1.08
          },
          "tp": {
            "type": "number",
            "example": 1.09
          },
          "deviation": {
            "type": "integer",
            "example": 20
          },
          "magic": {
            "type": "integer",
            "example": 10001
          },
          "comment": {
            "type": "string",
            "example": "Raw trade request"
          },
          "type_time": {
            "type": "integer",
            "example": 0
          },
          "type_filling": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "OrderCalcMarginRequest": {
        "type": "object",
        "required": [
          "action",
          "symbol",
          "volume",
          "price"
        ],
        "properties": {
          "action": {
            "type": "string",
            "description": "BUY or SELL (or integer code 0/1)",
            "example": "BUY"
          },
          "symbol": {
            "type": "string",
            "example": "EURUSD"
          },
          "volume": {
            "type": "number",
            "example": 1.0
          },
          "price": {
            "type": "number",
            "example": 1.085
          }
        }
      },
      "OrderCalcProfitRequest": {
        "type": "object",
        "required": [
          "action",
          "symbol",
          "volume",
          "price_open",
          "price_close"
        ],
        "properties": {
          "action": {
            "type": "string",
            "description": "BUY or SELL (or integer code 0/1)",
            "example": "BUY"
          },
          "symbol": {
            "type": "string",
            "example": "EURUSD"
          },
          "volume": {
            "type": "number",
            "example": 1.0
          },
          "price_open": {
            "type": "number",
            "example": 1.08
          },
          "price_close": {
            "type": "number",
            "example": 1.085
          }
        }
      }
    }
  }
}